tty: split tty_open to a separate file
Split the function tty_open() to a separate file.
